`Removed CommunityMenu, CreateMenu, and GenreMenu components and updated imports in header.tsx to point to new locations`
