+++ src/components/header/styles/header.module.css
@@ -1,0 +1,149 @@
+.header {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 50;
+  background: rgba(255, 255, 255, 0.2);
+  backdrop-filter: blur(16px);
+  padding: 0.75rem 1rem;
+}
+
+@media (min-width: 768px) {
+  .header {
+    padding: 0.75rem 1.5rem;
+  }
+}
+
+@media (min-width: 1024px) {
+  .header {
+    padding: 0.75rem 2rem;
+  }
+}
+
+.buttonText {
+  font-size: 1.125rem;
+}
+
+.menuText {
+  font-size: 1.125rem;
+}
+
+.exploraText {
+  font-size: 1.25rem;
+}
+
+.menuButton {
+  color: #187A25;
+  background: transparent;
+  padding: 0.5rem 0.75rem;
+  border-radius: 0.375rem;
+  transition: color 0.2s;
+  display: flex;
+  align-items: center;
+  gap: 0.25rem;
+}
+
+.menuButton:hover {
+  color: #4CD23D;
+}
+
+.chevronIcon {
+  height: 1rem;
+  width: 1rem;
+  transition: transform 0.2s;
+}
+
+.chevronIconOpen {
+  transform: rotate(180deg);
+}
+
+.dropdownContainer {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  margin-top: 0;
+  width: 550px;
+  background: white;
+  border: 1px solid #165C1E;
+  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1);
+  z-index: 50;
+  border-radius: 0.375rem;
+  overflow: hidden;
+}
+
+.dropdownHeader {
+  padding: 1rem;
+  border-bottom: 1px solid #e5e7eb;
+  background: #f9fafb;
+}
+
+.dropdownTitle {
+  font-size: 1.125rem;
+  font-weight: bold;
+  color: #187A25;
+}
+
+.dropdownGrid {
+  padding: 1rem;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 1rem;
+  min-height: 200px;
+}
+
+.dropdownLink {
+  display: block;
+  padding: 0.75rem 0.5rem;
+  color: #187A25;
+  border-radius: 0.25rem;
+  transition: all 0.2s;
+  font-size: 0.875rem;
+  font-weight: 500;
+  text-align: center;
+  line-height: 1.625;
+  word-break: break-word;
+  max-width: 100%;
+}
+
+.dropdownLink:hover {
+  background: rgba(76, 210, 61, 0.1);
+  color: #4CD23D;
+}
+
+.createDropdownContainer {
+  position: absolute;
+  top: 100%;
+  right: 0;
+  margin-top: 0.25rem;
+  width: 256px;
+  background: white;
+  border: 1px solid #165C1E;
+  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1);
+  z-index: 50;
+  border-radius: 0.375rem;
+  overflow: hidden;
+}
+
+.createDropdownList {
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+}
+
+.createDropdownLink {
+  display: block;
+  padding: 0.5rem 0.75rem;
+  color: #187A25;
+  border-radius: 0.25rem;
+  transition: all 0.2s;
+  font-size: 0.875rem;
+  font-weight: 500;
+  text-align: left;
+}
+
+.createDropdownLink:hover {
+  background: rgba(76, 210, 61, 0.1);
+  color: #4CD23D;
+}
